refactor: registration app re estructure
