+++ backend/cmd/config/dependencies.go
@@ -3,7 +3,7 @@ package config
 import (
 	"database/sql"
 
-	"hackathon-pvc-backend/internal/registration/application/services"
+	"hackathon-pvc-backend/internal/registration/app"
 	repo "hackathon-pvc-backend/internal/registration/infrastructure/persistance/sql"
 	"hackathon-pvc-backend/internal/registration/infrastructure/rest"
 )
@@ -14,12 +14,12 @@ type Dependencies struct {
 
 func WireDependencies(db *sql.DB) *Dependencies {
 	registrationRepository := repo.NewSqlRegistrationRepository(db)
-	// roleRepository := repo.NewSqlRoleRepository(db)
-	// technologyRepository := repo.NewSqlTechnologyRepository(db)
+	roleRepository := repo.NewSqlRoleRepository(db)
+	technologyRepository := repo.NewSqlTechnologyRepository(db)
 
-	registrationService := services.NewRegistrationService(registrationRepository)
-	// roleService := services.NewRoleService(roleRepository)
-	// technologyService := services.NewTechnologyService(technologyRepository)
+	roleService := app.NewRoleService(roleRepository)
+	technologyService := app.NewTechnologyService(technologyRepository)
+	registrationService := app.NewRegistrationService(registrationRepository, roleService, technologyService)
 
 	registrationHandler := rest.NewRegistrationHandler(registrationService)
