+++ backend/internal/registration/app/role.go
@@ -1,4 +1,4 @@
-package services
+package app
 
 import (
 	"context"
